requirements.txt
faker
pytest
pony
allure-pytest
selenium
helium
pytest-ordering
pytest-rerunfailures
requests
redis
pymysql
click